+++ common/src/main/java/nl/tudelft/wdm/group1/common/topic/OrdersTopics.java
@@ -1,4 +1,4 @@
-package nl.tudelft.wdm.group1.common.topics;
+package nl.tudelft.wdm.group1.common.topic;
 
 public class OrdersTopics {
     public static final String ORDER_CHECKED_OUT = "orderCheckedOut";
