+++ common/src/main/java/nl/tudelft/wdm/group1/common/topic/PaymentsTopics.java
@@ -1,4 +1,4 @@
-package nl.tudelft.wdm.group1.common.topics;
+package nl.tudelft.wdm.group1.common.topic;
 
 public class PaymentsTopics {
     public static final String PAYMENT_CREATED = "paymentCreated";
