+++ service/orders/src/main/java/nl/tudelft/wdm/group1/orders/events/Producer.java
@@ -1,7 +1,7 @@
 package nl.tudelft.wdm.group1.orders.events;
 
 import nl.tudelft.wdm.group1.common.model.Order;
-import nl.tudelft.wdm.group1.common.topics.OrdersTopics;
+import nl.tudelft.wdm.group1.common.topic.OrdersTopics;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
