+++ service/payments/src/main/java/nl/tudelft/wdm/group1/payments/events/Consumer.java
@@ -2,8 +2,8 @@ package nl.tudelft.wdm.group1.payments.events;
 
 import nl.tudelft.wdm.group1.common.exception.ResourceNotFoundException;
 import nl.tudelft.wdm.group1.common.model.Payment;
-import nl.tudelft.wdm.group1.common.topics.PaymentsTopics;
-import nl.tudelft.wdm.group1.common.topics.UsersTopics;
+import nl.tudelft.wdm.group1.common.topic.PaymentsTopics;
+import nl.tudelft.wdm.group1.common.topic.UsersTopics;
 import nl.tudelft.wdm.group1.payments.PaymentRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
