+++ service/rest/src/test/java/nl/tudelft/wdm/group1/rest/RestAddUserTest.java
@@ -4,7 +4,7 @@ import nl.tudelft.wdm.group1.common.*;
 import nl.tudelft.wdm.group1.common.exception.ResourceNotFoundException;
 import nl.tudelft.wdm.group1.common.model.User;
 import nl.tudelft.wdm.group1.common.payload.UserCreatePayload;
-import nl.tudelft.wdm.group1.common.topics.RestTopics;
+import nl.tudelft.wdm.group1.common.topic.RestTopics;
 import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
