+++ service/rest/src/test/java/nl/tudelft/wdm/group1/rest/WdmKafkaTestHelpers.java
@@ -1,6 +1,6 @@
 package nl.tudelft.wdm.group1.rest;
 
-import nl.tudelft.wdm.group1.common.topics.RestTopics;
+import nl.tudelft.wdm.group1.common.topic.RestTopics;
 import org.apache.kafka.clients.consumer.*;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
