+++ service/stock/src/main/java/nl/tudelft/wdm/group1/stock/events/Consumer.java
@@ -6,8 +6,8 @@ import nl.tudelft.wdm.group1.common.exception.ResourceNotFoundException;
 import nl.tudelft.wdm.group1.common.model.Order;
 import nl.tudelft.wdm.group1.common.model.OrderStatus;
 import nl.tudelft.wdm.group1.common.model.StockItem;
-import nl.tudelft.wdm.group1.common.topics.OrdersTopics;
-import nl.tudelft.wdm.group1.common.topics.StockTopics;
+import nl.tudelft.wdm.group1.common.topic.OrdersTopics;
+import nl.tudelft.wdm.group1.common.topic.StockTopics;
 import nl.tudelft.wdm.group1.stock.StockItemRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
