+++ service/users/src/main/java/nl/tudelft/wdm/group1/users/events/Consumer.java
@@ -2,7 +2,7 @@ package nl.tudelft.wdm.group1.users.events;
 
 import nl.tudelft.wdm.group1.common.exception.ResourceNotFoundException;
 import nl.tudelft.wdm.group1.common.model.User;
-import nl.tudelft.wdm.group1.common.topics.UsersTopics;
+import nl.tudelft.wdm.group1.common.topic.UsersTopics;
 import nl.tudelft.wdm.group1.users.UserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
