+++ service/users/src/main/java/nl/tudelft/wdm/group1/users/events/PaymentConsumer.java
@@ -5,7 +5,7 @@ import nl.tudelft.wdm.group1.common.exception.InsufficientCreditException;
 import nl.tudelft.wdm.group1.common.exception.ResourceNotFoundException;
 import nl.tudelft.wdm.group1.common.model.Payment;
 import nl.tudelft.wdm.group1.common.model.User;
-import nl.tudelft.wdm.group1.common.topics.PaymentsTopics;
+import nl.tudelft.wdm.group1.common.topic.PaymentsTopics;
 import nl.tudelft.wdm.group1.users.UserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
