+++ service/users/src/main/java/nl/tudelft/wdm/group1/users/events/PaymentProducer.java
@@ -1,6 +1,6 @@
 package nl.tudelft.wdm.group1.users.events;
 
-import nl.tudelft.wdm.group1.common.topics.UsersTopics;
+import nl.tudelft.wdm.group1.common.topic.UsersTopics;
 import nl.tudelft.wdm.group1.common.model.Payment;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
