+++ service/users/src/main/java/nl/tudelft/wdm/group1/users/events/Producer.java
@@ -1,7 +1,7 @@
 package nl.tudelft.wdm.group1.users.events;
 
 import nl.tudelft.wdm.group1.common.model.User;
-import nl.tudelft.wdm.group1.common.topics.UsersTopics;
+import nl.tudelft.wdm.group1.common.topic.UsersTopics;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
